Testing for sequence number issues
We hit an issue where deleting one batch of mails and then deleting
another fails because the first deletion (plus expunge) causes the
second set of sequence numbers to be wrong.

Here we tweak our dovecot launch so sequence numbers don't start at 1
(for future testing with UIDs) and added a test that does exactly that,
and fails.